feat(typescript): add typings to `withGroupProfile` wrapper
